+++ lib/src/commands/commands.dart
@@ -1,3 +1,7 @@
+export 'build/build_command.dart';
+export 'clean/clean_command.dart';
 export 'create/create_command.dart';
 export 'doctor/doctor_command.dart';
 export 'generate/generate_command.dart';
+export 'pub/pub_command.dart';
+export 'sdk/sdk_command.dart';